Use Logger block form calls
When blocks are used, and the set logging
level of the Logger object is above that of
the call, the block will not be executed.

This prevents string allocations in log
messages that would not be retained anyway,
which can be especially significant when
calling `inspect` on large-ish objects.

For rationale, see http://hawkins.io/2013/08/using-the-ruby-logger/
and https://stackoverflow.com/questions/30144317