Add autoloadability via Bundler.require
Bundler can automatically load gems if there is a conventionally named.
First, it will try to load a file matching the name of the gem:

  https://github.com/bundler/bundler/blob/v2.0.2/lib/bundler/runtime.rb#L76-L87

Then, it will rely on Bundler's naming conventions:

  https://guides.rubygems.org/name-your-gem/

This commit adds a file named `launchdarkly-server-sdk.rb` that simply
requires `ldclient-rb.rb`, allowing for this feature to be used.

This allows Rails to automatically load the gem via Bundler.require:

  https://github.com/rails/rails/blob/v5.2.3/railties/lib/rails/generators/rails/app/templates/config/application.rb.tt#L20-L22