Fix JSON serialization failure (#237)
When we introduced models for the flag and segment data, we added
to_json methods which proxy to the underlying to_json method provided by
the `json` gem.

We defined the method with a variadic parameter, but we failed to unpack
it when passing it on to the underlying implementation. This resulted in
a serialization failure which prevented the redis data store from
initializing.