test: Annotate and disable flaky tests in JRuby (#247)
Webrick seems to struggle when running on JRuby. I suspect this because
of the global lock in standard ruby but I don't know for sure.

Either way, having tests fail randomly is no way to live so I am
disabling them in JRuby.

I don't think this is a huge concern because

- The same tests are run in other rubies, so we know the coded logic is
sound.

- The failures don't originate from our code -- rather, they seem to
come from the mock server, so no important code is being left untested.