fix: Add additional payload filter key validation
Previously, we were allowing any non-empty string value to be provided
as a payload filter key. However, customers can only create a filter key
with a subset of characters.

In an effort to warn them earlier but potentially invalid
configurations, we are adding some basic key validation work as part of
the start up sequence.