fix: Add additional payload filter key validation (#295)
Previously, we were allowing any non-empty string value to be provided as a
payload filter key. However, customers can only create a filter key with a
subset of characters within the LaunchDarkly UI.

In an effort to warn customers earlier about potentially invalid
configurations, we are adding basic key validation as part of the start up
sequence.