chore: Allow OS to assign port for unit tests (#336)
Previously we started looking for testing ports starting at 5000. If the
port was in use, we would incrementally search upwards until we found a
port we could use.

A recent change in the GH Windows image resulted in a permission error
being raised before we could find a port to use:

```
Failure/Error: WEBrick::HTTPServer.new(base_opts)

Errno::EACCES:
Permission denied - bind(2) for 127.0.0.1:50018
```

To address this, we are modifying the tests to set a port of 0, which
will allow the OS to assign an available port for us. This removes the
need for this silly retry logic and hard coded port numbers.