fix: Explicitly require openssl gem (#334)
OpenSSL 3.6.0 introduced a change in behavior which prevents the SDK
from initializing. To learn more, refer to [this github ticket][1].

To mitigate this, we explicitly depend on the openssl gem instead of
loading the default version shipped with Ruby.

This SDK still advertises Ruby 3.1+ support, which should require a
minimum threshold of openssl-v3.0+. However, Ruby 3.1 is EOL and the
openssl maintainer only implemented fixes in openssl-v3.1+. Thus, we set
the minimum to openssl-v3.1.

[1]: https://github.com/ruby/openssl/issues/949

fixes #333